// src/models/mufg.ts
export interface MufgTrade {
    File_Type: string;
    Fund: string;
    Transaction_Event: string;
    Transaction_Type: string;
    Security_ID_ISIN: string;
    Security_ID_CUSIP: string;
    Security_ID_SEDOL: string;
    Security_ID_Bloomberg: string;
    Security_ID_Reuters: string;
    Security_ID_UGC: string;
    Security_Description: string;
    Trade_ID_Client: string;
    Quantity: string;
    Original_Face: string;
    Price: string;
    Accrued_Interest: string;
    Net_Money_Settlement: string;
    Currency_Settlement: string;
    Currency_Investment: string;
    Trade_Date: string;
    Settle_Date: string;
    Expiration_Date: string;
    Strike: string;
    Put_Call: string;
    Custodian_Account_Client: string;
    Custodian_Account_UGC: string;
    Broker_Client: string;
    Broker_UGC: string;
    Fund_Structure: string;
    Strategy_Client: string;
    Strategy_UGC: string;
    Comments: string;
    Order_Commission: string;
    Trader_Client: string;
    Trader_UGC: string;
    Manager_Client: string;
    Manager_UGC: string;
    Analyst_Client: string;
    Analyst_UGC: string;
    Industry_Client: string;
    Industry_UGC: string;
    Underlying_ISIN: string;
    Underlying_CUSIP: string;
    Underlying_Sedol: string;
    Underlying_BBG: string;
    Underlying_RIC: string;
    Trade_Expense_1_Net: string;
    Trade_Expense_1_Type: string;
    Trade_Expense_2_Net: string;
    Trade_Expense_2_Type: string;
    Trade_Expense_3_Net: string;
    Trade_Expense_3_Type: string;
    Commission_NetTrade: string;
    CFD_Flag: string;
    Security_Country: string;
    Closing_Lot_ID: string;
    Secondary_Client_Trade_ID: string;
    Net_Money_Trade: string;
    Is_Factor: string;
    Underlying_UGC: string;
    Underlying_Desc: string;
    Underlying_Country: string;
    Location: string;
  }